# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: APIs-Graduation
  version: 1.0.0
paths:
...
  /api/search/doctors/search:
    get:
      tags:
        - 'Search For A Doctor'
      summary: 'Search for doctors by name or specialization'
      operationId: f5288f5ae4cbd5c1f0cc80d931ba11af
      parameters:
        -
          name: query
          in: query
          description: 'Search query for doctor first name or specialization name'
          required: true
          schema:
            type: string
            example: Cardiology
      responses:
        200:
          description: 'Doctors retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  doctors:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        first_name:
                          type: string
                          example: John
                        last_name:
                          type: string
                          example: Doe
                        full_name:
                          type: string
                          example: 'John Doe'
                        specialization:
                          type: string
                          example: Cardiology
                          nullable: true
                        experience:
                          type: integer
                          example: 10
                          nullable: true
                        rating:
                          type: number
                          format: float
                          example: 4.5
                        about:
                          type: string
                          example: 'Experienced cardiologist'
                          nullable: true
                        salary:
                          type: number
                          example: 5000
                          nullable: true
                        schedule:
                          $ref: '#/components/schemas/DoctorSchedule'
                        image_name:
                          type: string
                          example: 'http://your-app-url/storage/images/doctor_img/doctor.jpg'
                          nullable: true
                      type: object
                  message:
                    type: string
                    example: 'Doctors retrieved successfully'
                  status:
                    type: integer
                    example: 200
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                  status:
                    type: integer
                    example: 401
                type: object
        403:
          description: 'Unauthorized action'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'You are not authorized to search doctors.'
                  status:
                    type: integer
                    example: 403
                type: object
        404:
          description: 'No doctors found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'No doctor found with this name'
                  status:
                    type: integer
                    example: 404
                type: object
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                  errors:
                    type: object
                  status:
                    type: integer
                    example: 422
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    DoctorSchedule:
      properties:
        Monday:
          type: array
          items:
            type: string
            example: '10:00 AM, 1:00PM, 4:00PM'
        Tuesday:
          type: array
          items:
            type: string
            example: '9:00 AM, 1:00PM, 4:00PM'
      type: object
...